requirements.txt
Flask==2.3.3
azure-ai-formrecognizer==3.3.0
azure-core==1.29.4
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
msal==1.24.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7